refactor the BarcodeFragment
